refactor: Split form into components and add validations
- Refactor PhoneAndBirthdayComponent to display phone and birthday fields in a row.
- Add validation for phone number (10 digits) and birthday (DD-MM-YYYY format, no future dates).
- Update styles for field alignment and error messages.
