+++ app/components/login/EnterPasswordComponent.jsx
@@ -75,6 +75,7 @@ const EnterPasswordComponent = ({ password, setPassword, textsLeng }) => {
 const styles = StyleSheet.create({
   container: {
     width: "90%",
+    marginTop: -15
   },
   label: {
     color: Colors.primary,
@@ -94,7 +95,7 @@ const styles = StyleSheet.create({
   },
   errorText: {
     color: "red",
-    fontSize: SizeConstants.texts - 3,
+    fontSize: SizeConstants.texts - 5,
     marginTop: 2,
     marginBottom: 10,
   },
